backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
alembic==1.13.3
pytesseract==0.3.13
Pillow==11.0.0
pdfplumber==0.11.4
PyMuPDF==1.23.8
redis==5.0.8
rapidocr-onnxruntime==1.3.24